Dynamic core can now run Win98 if the blocks are not cached.  This shows that the instructions are handled correctly but that clearing the cache when memory containing code is overwritten still needs some work.
git-svn-id: https://svn.code.sf.net/p/jdosbox/code/trunk@447 4cb9e868-4783-4b93-94d1-9e8e689819c6

+++ src/jdos/cpu/Paging.java
@@ -49,7 +49,7 @@ public class Paging extends Module_base {
         }
 
         public /*Bitu*/long readd(/*PhysPt*/long addr) {
-            return readb(addr + 0) | (readb(addr + 1) << 8) | (readb(addr + 2) << 16) | (readb(addr + 3) << 24);
+            return readb(addr + 0) | (readb(addr + 1) << 8) | (readb(addr + 2) << 16) | ((long)readb(addr + 3) << 24);
         }
 
         public void writeb(/*PhysPt*/long addr,/*Bitu*/int val) {
